+++ src/components/Combobox/Combobox.scss
@@ -2,8 +2,6 @@
 @import "../../styles/typography.scss";
 @import "../../styles/states.scss";
 
-$icon-margin: 4px;
-
 .combobox--wrapper {
   height: 100%;
   display: flex;
@@ -57,53 +55,5 @@ $icon-margin: 4px;
   &-list {
     flex: 1 1 auto;
     padding: 8px 4px;
-
-    .category-label {
-      @include theme-prop(color, placeholder-color);
-      padding-bottom: 4px;
-    }
-
-    .combobox-option {
-      padding: 8px;
-      border-radius: 4px;
-      display: flex;
-      align-items: center;
-      overflow: hidden;
-      cursor: pointer;
-      user-select: none;
-
-      &.disabled {
-        cursor: not-allowed;
-        @include disabled;
-      }
-
-      &.selected {
-        @include theme-prop(background-color, primary-selected-color);
-      }
-
-      &:hover:not(.disabled):not(.selected),
-      &.active {
-        @include theme-prop(color, primary-text-color);
-        @include theme-prop(background-color, primary-background-hover-color);
-      }
-
-      &.active {
-        position: relative;
-        &.active-outline {
-          @include focus-style-css();
-        }
-      }
-
-      .option-label {
-        flex: 1 1 auto;
-      }
-
-      .option-icon {
-        flex: 0 0 auto;
-        &.left {
-          margin-right: $icon-margin;
-        }
-      }
-    }
   }
 }
